mova on message action to another module
